Change package and add license
