+++ ultralight-java-base/src/main/java/com/labymedia/ultralight/plugin/render/UltralightGPUState.java
@@ -1,7 +1,27 @@
-package net.janrupf.ultralight.plugin.render;
+/*
+ * Ultralight Java - Java wrapper for the Ultralight web engine
+ * Copyright (C) 2020 LabyMedia and contributors
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
 
-import net.janrupf.ultralight.annotation.NativeType;
-import net.janrupf.ultralight.math.IntRect;
+package com.labymedia.ultralight.plugin.render;
+
+import com.labymedia.ultralight.annotation.NativeType;
+import com.labymedia.ultralight.math.IntRect;
+
 
 @NativeType("ultralight::GPUState")
 public class UltralightGPUState {
@@ -28,7 +48,8 @@ public class UltralightGPUState {
 
     private final @NativeType("bool") boolean enableScissor;
 
-    private final @NativeType("IntRect") IntRect scissorRect;
+    private final @NativeType("IntRect")
+    IntRect scissorRect;
 
     public UltralightGPUState(long viewportWidth,
                               long viewportHeight,
