--- ultralight-java-base/src/main/java/net/janrupf/ultralight/plugin/render/UltralightCommandList.java
@@ -1,8 +1,0 @@
-package net.janrupf.ultralight.plugin.render;
-
-import net.janrupf.ultralight.annotation.NativeType;
-
-import java.util.ArrayList;
-
-@NativeType("ultralight::CommandList")
-public class UltralightCommandList extends ArrayList<UltralightCommand> {}
