--- ultralight-java-base/src/main/java/net/janrupf/ultralight/plugin/render/UltralightIndexBuffer.java
@@ -1,9 +1,0 @@
-package net.janrupf.ultralight.plugin.render;
-
-import net.janrupf.ultralight.annotation.NativeType;
-
-import java.util.ArrayList;
-
-@NativeType("ultralight::IndexBuffer")
-public class UltralightIndexBuffer extends ArrayList<Short> {
-}
