--- ultralight-java-base/src/main/java/net/janrupf/ultralight/plugin/render/UltralightRenderBuffer.java
@@ -1,7 +1,0 @@
-package net.janrupf.ultralight.plugin.render;
-
-import net.janrupf.ultralight.annotation.NativeType;
-
-@NativeType("ultralight::RenderBuffer")
-public class UltralightRenderBuffer {
-}
