--- ultralight-java-base/src/main/java/net/janrupf/ultralight/plugin/render/UltralightVertexBufferFormat.java
@@ -1,9 +1,0 @@
-package net.janrupf.ultralight.plugin.render;
-
-import net.janrupf.ultralight.annotation.NativeType;
-
-@NativeType("ultralight::VertexBufferFormat")
-public enum UltralightVertexBufferFormat {
-    FORMAT_2F_4UB_2F,
-    FORMAT_2F_4UB_2F_2F_28F;
-}
